Built properties
In some cases, the properties are not defined in the junction/interface in the excel file, for that reason I suggested to modify 'i' to "i-1". I have found that in some cases the mobility goes to infinite at the interface and pdepe claims that is not a good initial value due to the mobility and tau values. What I did is to let constant tau (as I commented in the scripts of 'surface_re_taun' taun= 1/sn(i)), and in build_device.m I put the mobility 'lin_graded'.
I have included Check_build_device_properties.m that includes the code to plot dev properties.
I used to run it before the end of build_device.m to check if the arrays of parameters are correct.